Add a few unit tests for DustTasks
They test the logic for getting a Dust template name from the file path.
There are two tests, for Unix and Windows respectively. The Unix one
passes, and the Windows one fails (deliberately).

The intention with the failing test is to illustrate a critical Windows
bug, which in effect makes it impossible to use the Dust plugin in a
Windows environment!

The bug is due to the fact that Windows uses backslashes for directory
separators, while the Dust plugin assumes forward slashes.

The next commit will make changes to the logic, to make the test pass and
fix the bug.